+++ packages/@headlessui-react/src/components/disclosure/disclosure.test.tsx
@@ -597,7 +597,7 @@ describe('Composition', () => {
           <Transition>
             <Debug name="Transition" fn={orderFn} />
             <Disclosure.Panel>
-              <Transition.Child>
+              <Transition.Child as="div">
                 <Debug name="Transition.Child" fn={orderFn} />
               </Transition.Child>
             </Disclosure.Panel>