+++ packages/@headlessui-react/src/components/transition/transition.tsx
@@ -49,6 +49,44 @@ function splitClasses(classes: string = '') {
   return classes.split(/\s+/).filter((className) => className.length > 1)
 }
 
+/**
+ * Check if we should forward the ref to the child element or not. This is to
+ * prevent crashes when the `as` prop is a Fragment _and_ the component just acts
+ * as a state container (aka, there is no actual transition happening).
+ *
+ * E.g.:
+ *
+ * ```tsx
+ * <Transition show={true}>
+ *   <Transition.Child enter="duration-100"><div>Child 1</div></Transition.Child>
+ *   <Transition.Child enter="duration-200"><div>Child 2</div></Transition.Child>
+ * </Transition>
+ * ```
+ *
+ * In this scenario, the child components are transitioning, but the
+ * `Transition` parent, which is a `Fragment`, is not. So we should not forward
+ * the ref to the `Fragment`.
+ */
+function shouldForwardRef<TTag extends ElementType = typeof DEFAULT_TRANSITION_CHILD_TAG>(
+  props: TransitionRootProps<TTag>
+) {
+  return (
+    // If we have any of the enter/leave classes
+    Boolean(
+      props.enter ||
+        props.enterFrom ||
+        props.enterTo ||
+        props.leave ||
+        props.leaveFrom ||
+        props.leaveTo
+    ) ||
+    // If the `as` prop is not a Fragment
+    (props.as ?? DEFAULT_TRANSITION_CHILD_TAG) !== Fragment ||
+    // If we have a single child, then we can forward the ref directly
+    React.Children.count(props.children) === 1
+  )
+}
+
 interface TransitionContextValues {
   show: boolean
   appear: boolean
@@ -264,7 +302,7 @@ function useNesting(done?: () => void, parent?: NestingContextValues) {
 
 // ---
 
-let DEFAULT_TRANSITION_CHILD_TAG = 'div' as const
+let DEFAULT_TRANSITION_CHILD_TAG = Fragment
 type TransitionChildRenderPropArg = MutableRefObject<HTMLDivElement>
 let TransitionChildRenderFeatures = RenderFeatures.RenderStrategy
 
@@ -292,7 +330,9 @@ function TransitionChildFn<TTag extends ElementType = typeof DEFAULT_TRANSITION_
     ...rest
   } = props as typeof props
   let container = useRef<HTMLElement | null>(null)
-  let transitionRef = useSyncRefs(container, ref)
+  let requiresRef = shouldForwardRef(props)
+
+  let transitionRef = useSyncRefs(...(requiresRef ? [container, ref] : ref === null ? [] : [ref]))
   let strategy = rest.unmount ?? true ? RenderStrategy.Unmount : RenderStrategy.Hidden
 
   let { show, appear, initial } = useTransitionContext()
@@ -342,10 +382,12 @@ function TransitionChildFn<TTag extends ElementType = typeof DEFAULT_TRANSITION_
   let ready = useServerHandoffComplete()
 
   useIsoMorphicEffect(() => {
+    if (!requiresRef) return
+
     if (ready && state === TreeStates.Visible && container.current === null) {
       throw new Error('Did you forget to passthrough the `ref` to the actual DOM node?')
     }
-  }, [container, state, ready])
+  }, [container, state, ready, requiresRef])
 
   // Skipping initial transition
   let skip = initial && !appear
@@ -495,7 +537,11 @@ function TransitionRootFn<TTag extends ElementType = typeof DEFAULT_TRANSITION_C
   // @ts-expect-error
   let { show, appear = false, unmount = true, ...theirProps } = props as typeof props
   let internalTransitionRef = useRef<HTMLElement | null>(null)
-  let transitionRef = useSyncRefs(internalTransitionRef, ref)
+  let requiresRef = shouldForwardRef(props)
+
+  let transitionRef = useSyncRefs(
+    ...(requiresRef ? [internalTransitionRef, ref] : ref === null ? [] : [ref])
+  )
 
   // The TransitionChild will also call this hook, and we have to make sure that we are ready.
   useServerHandoffComplete()